# Sprint 39: Critical fixes for state management and performance