… TEAM-CONTEXT.md
… memory-bank/api-docs/sprints/sprint11.md
… memory-bank/sprints/sprint35/IMAGE-ANALYSIS-CRITICAL-EXAMINATION.md
… memory-bank/sprints/sprint35/IMAGE-CONTEXT-INTEGRATION-PLAN.md
… memory-bank/sprints/sprint38/AUTOFIXER-EXPORT-PATTERN-FIX.md
… memory-bank/sprints/sprint38/AUTOFIXER-UI-PROBLEMS-ANALYSIS.md
… memory-bank/sprints/sprint38/IMAGE-PERSISTENCE-AND-ANALYSIS-FIX.md
… memory-bank/sprints/sprint38/IMAGE-PERSISTENCE-COMPLETE.md
… memory-bank/sprints/sprint39/ADD-SCENE-STATE-UPDATE-DEBUG.md
… memory-bank/sprints/sprint39/CONTEXT-OPTIMIZATION-PLAN.md
… memory-bank/sprints/sprint39/CONTEXT-SYSTEM-DATABASE-FIX.md
… memory-bank/sprints/sprint39/CREATE-SCENE-FROM-IMAGE-RESPONSE-ISSUE.md
… memory-bank/sprints/sprint39/CRITICAL-SCENE-NUMBERING-BUG.md
… memory-bank/sprints/sprint39/IMAGE-SYSTEM-PHASE1-80-COMPLETE.md
… memory-bank/sprints/sprint39/IMAGE-SYSTEM-PHASE2-80-COMPLETE.md
… memory-bank/sprints/sprint39/IMMEDIATE-FIX-ADD-SCENE.md
… memory-bank/sprints/sprint39/ORCHESTRATOR-PREFERENCE-CLEANUP.md
… memory-bank/sprints/sprint39/PREFERENCE-EXTRACTOR-CONFIG-REFACTOR.md
… memory-bank/sprints/sprint39/QUICK-WINS-IMMEDIATE.md
… memory-bank/sprints/sprint39/README.md
… memory-bank/sprints/sprint39/SCENE-NAMING-CONSISTENCY-ISSUES.md
… memory-bank/sprints/sprint39/STATE-MANAGEMENT-ADD-SCENE-DURATION-FIX.md
… memory-bank/sprints/sprint39/STATE-MANAGEMENT-SIMPLIFICATION-COMPLETE.md
… memory-bank/sprints/sprint39/state-management-scene-operations-fix.md
… src/server/services/brain/contextBuilder.service.ts
… src/server/services/brain/orchestrator.ts
… src/server/services/brain/preferenceExtractor.service.ts