## Summary

This PR contains critical fixes from Sprint 39 that resolve multiple production issues and sets up performance optimizations.

## Key Fixes

### 1. 🐛 Fixed Autofixer UI Errors
- Fixed `hasSceneError` undefined error
- Updated to use `sceneErrors` Map for tracking multiple scene errors
- Added scene existence validation to prevent stale UI

### 2. 🔧 Fixed Context Building Errors  
- Fixed `imageAnalysisPromise` undefined error by removing async image analysis
- Fixed `userContext` is not defined error in buildContextPacket
- Fixed `dbPreferences.forEach` error (was object, not array)

### 3. 🎬 Fixed Scene Operations
- **addScene**: Now properly transforms database scenes to InputProps format with timeline calculations
- **createSceneFromImage**: Added workaround for missing scene ID in response
- **editScene**: Already working, maintained compatibility
- **changeDuration**: Fixed with targetSceneId support

### 4. 📚 Sprint 39 Documentation
- Context optimization plan targeting 80-90% performance improvement
- Quick wins for immediate performance gains
- Scene naming consistency issues and solutions
- Comprehensive debugging guides

## Testing

All scene operations tested:
- ✅ Add scene (text-based)
- ✅ Add scene from image (with workaround)
- ✅ Edit scene
- ✅ Change duration
- ✅ Delete scene
- ✅ Autofixer UI

## Performance Impact

With the documented quick wins:
- Simple edits: 70s → 5s (93% improvement potential)
- Add scene: Currently working, will be 10s with optimizations
- Autofix: 70s → 3s (96% improvement potential)

## What's Next

The Sprint 39 documentation includes:
- `/memory-bank/sprints/sprint39/CONTEXT-OPTIMIZATION-PLAN.md` - Full 4-week optimization strategy
- `/memory-bank/sprints/sprint39/QUICK-WINS-IMMEDIATE.md` - 5 fixes that can be implemented TODAY
- `/memory-bank/sprints/sprint39/SCENE-NAMING-CONSISTENCY-ISSUES.md` - Solutions for scene naming problems

## Notes

- Merged latest dev changes and resolved conflicts
- Kept our improved state management approach over the old refetch pattern
- All existing functionality preserved while adding fixes